# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: NVR API
  version: 0.1.0
  description: Contract for Go (oapi-codegen) and packages/api-client (openapi-typescript).
servers:
  - url: /api/v1
paths:
...
  /storage/gdrive/status:
    get:
      operationId: getGDriveStatus
      tags: [storage]
      summary: Google Drive connection status
      description: Authenticated users. Returns OAuth configuration and connection status (never tokens).
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GDriveStatus"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /storage/gdrive/configuration:
    put:
      operationId: putGDriveConfiguration
      tags: [storage]
      summary: Save Google Drive OAuth configuration
      description: |
        Admin only. Stores the Google OAuth client configuration used to connect
        Drive. The client secret is encrypted at rest with NVR_SECRETS_KEY.
        Saving a new configuration clears any existing Drive connection.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GDriveConfigurationRequest"
      responses:
        "200":
          description: Configuration saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GDriveStatus"
        "400":
          description: Invalid or insecure configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /storage/gdrive/connect:
    post:
      operationId: postGDriveConnect
      tags: [storage]
      summary: Begin Google Drive OAuth connect
      description: Admin only. Returns the Google authorization URL to open in a browser.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Authorization URL ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GDriveConnectResponse"
        "400":
          description: Google OAuth not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /storage/gdrive/callback:
    get:
      operationId: getGDriveCallback
      tags: [storage]
      summary: Google Drive OAuth callback
      description: |
        Public OAuth redirect target. Exchanges the authorization code, stores tokens,
        and redirects (302) to the settings UI.
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
        - name: error_description
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to settings UI after connect success or failure
          headers:
            Location:
              description: Frontend settings URL with gdrive status query params
              schema:
                type: string
  /storage/gdrive/disconnect:
    delete:
      operationId: deleteGDriveDisconnect
      tags: [storage]
      summary: Disconnect Google Drive
      description: Admin only. Clears stored tokens and best-effort revokes with Google.
      security:
        - cookieAuth: []
      responses:
        "204":
          description: Disconnected
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /storage/gdrive/archive:
    post:
      operationId: postGDriveArchive
      tags: [storage]
      summary: Archive recordings to Google Drive
      description: |
        Admin only. Uploads a batch of unarchived local recordings to the connected
        Google Drive account (same work as the recurring background archive job).
      security:
        - cookieAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GDriveArchiveRequest"
      responses:
        "200":
          description: Archive batch result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GDriveArchiveResponse"
        "400":
          description: Drive not connected or invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: An archive pass is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nvr_session
      description: Session cookie issued on login or setup.
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    GDriveStatus:
      type: object
      required: [configured, connected]
      properties:
        configured:
          type: boolean
          description: True when Google OAuth client details are saved in the dashboard or supplied by environment, and NVR_SECRETS_KEY is set
        connected:
          type: boolean
          description: True when stored credentials can be decrypted
        connectionError:
          type: string
          description: Recoverable stored-credential problem; reconnecting resolves it
        accountEmail:
          type: string
          description: Connected Google account email; returned only to administrators
        connectedAt:
          type: string
          format: date-time
        folderId:
          type: string
    GDriveConfigurationRequest:
      type: object
      required: [clientId, clientSecret, redirectUrl]
      properties:
        clientId:
          type: string
          minLength: 1
          description: OAuth web client ID from Google Cloud Console
        clientSecret:
          type: string
          minLength: 1
          writeOnly: true
          description: OAuth web client secret; encrypted at rest and never returned
        redirectUrl:
          type: string
          format: uri
          description: Authorized OAuth redirect URI configured in Google Cloud Console
    GDriveConnectResponse:
      type: object
      required: [authorizationUrl]
      properties:
        authorizationUrl:
          type: string
          description: Google OAuth consent URL
    GDriveArchiveRequest:
      type: object
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 50
          description: Max segments to archive in this batch (default 50)
    GDriveArchiveResponse:
      type: object
      required: [uploaded, failed, skipped]
      properties:
        uploaded:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
...